+++ src/index.js
@@ -1,15 +1,16 @@
 //Get dependencies
-const impl = require('./core');
+const coreImpl = require('./core');
+const collectionImpl = require('./collection');
 
 // registers the extension on a cytoscape lib ref
 let register = function( cytoscape ){
   if( !cytoscape ){ return; } // can't register if cytoscape unspecified
 
   // register with cytoscape.js
-  cytoscape( 'core', 'popper', impl.core );  //Cytoscape Core
-  cytoscape( 'collection', 'popper', impl.collection); //Cytoscape Collections
-  cytoscape( 'core', 'popperRef', impl.coreRef );  //Cytoscape Core for References 
-  cytoscape( 'collection', 'popperRef', impl.collectionRef); //Cytoscape Collections for References
+  cytoscape( 'core', 'popper', coreImpl.core );  //Cytoscape Core
+  cytoscape( 'collection', 'popper', collectionImpl.collection); //Cytoscape Collections
+  cytoscape( 'core', 'popperRef', coreImpl.coreRef );  //Cytoscape Core for References 
+  cytoscape( 'collection', 'popperRef', collectionImpl.collectionRef); //Cytoscape Collections for References
 
 };
